aboo glm
Utbilden just nu är 200x200

+++ Raytracer/Raytracer/ColorDbl.h
@@ -8,10 +8,10 @@ class ColorDbl {
 public:
 	ColorDbl() : R(0), G(0), B(0) { };
 	//Constructor
-	ColorDbl(double r, double g, double b) : R(r), G(g), B(b) {};
+	ColorDbl(float r, float g, float b) : R(r), G(g), B(b) {};
 
-	ColorDbl operator*(double const num) {
-		double x = num;
+	ColorDbl operator*(float const num) {
+		float x = num;
 		if (x > 1) {
 			x = 1;
 		}
@@ -21,8 +21,8 @@ public:
 		return ColorDbl(R*x, G*x, B*x);
 	};
 
-	ColorDbl operator/(double const num) {
-		double x = num;
+	ColorDbl operator/(float const num) {
+		float x = num;
 		if (x > 1) {
 			x = 1;
 		}
@@ -33,24 +33,24 @@ public:
 	};
 
 	ColorDbl operator*(ColorDbl &c1) {
-		c1.R = c1.R / 255;
-		c1.G = c1.G / 255;
-		c1.B = c1.B / 255;
+		c1.R = c1.R / 255.0f;
+		c1.G = c1.G / 255.0f;
+		c1.B = c1.B / 255.0f;
 		return ColorDbl(this->R*c1.R, this->G*c1.G, this->B*c1.B);
 	};
 
 	ColorDbl operator+(ColorDbl c1) {
-		int x = (R + c1.R)/2;
-		int y = (G + c1.G)/2;
-		int z = (B + c1.B)/2;
+		float x = (R + c1.R) / 2.0f;
+		float y = (G + c1.G) / 2.0f;
+		float z = (B + c1.B) / 2.0f;
 		return ColorDbl(x, y, z);
 	};
 
 	//ColorDbl shading(Ray &importance, Triangle hitTri, Scene s);
 
-	double R;
-	double G;
-	double B;
+	float R;
+	float G;
+	float B;
 private:
 
 	//friend std::
